Introduction of SimpleContract and related services, moved to fully async implemenation
All Attrubutes required by an algo trading platfofrm are included in SimpleContact, a json Redis OM entity.
Communication with TWS now fully asynchronous using java Future with a TimeOut
ExecHandler interface introduced to Allow Callbacks to TWS client apps
An example client is included. SimpleContactController and associated Service. This allows us to create a Order based on the minimum set of criteria needed for the Contract definition. The service returns the SimpleContract used which is stored in Redis
Some classes like OrderManagementService which is used by the new SimpleContractService are modified to reflect use of ExecHandler